# Optimize DSL