@dennwc 2019-03-14
Merging this one, since it's blocking driver update.

@bzz Let me know if you have any comments.